Add map modal to dashboard
