+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - Updated Plan Wizard to allow selecting Organization boundaries
  - Added map page
+ - Added map modal to Risks on Dashboard page
  - Added County & ClimateAssessmentRegion models/fixtures
  - Added County GeoJSON API endpoint
  - Added mini map to dashboard
