+++ src/angular/planit/src/assets/fonts/fontello/config.json
@@ -1765,6 +1765,20 @@
       ]
     },
     {
+      "uid": "ca947770aeef0a4cbd42a5b41e7a3e6a",
+      "css": "map-marked-alt",
+      "code": 59543,
+      "src": "custom_icons",
+      "selected": true,
+      "svg": {
+        "path": "M562.5 0C426.6 0 316.4 110.2 316.4 246.1 316.4 356 477.2 556.3 538.9 628.9 551.3 643.7 573.7 643.7 586.1 628.9 647.8 556.3 808.6 356 808.6 246.1 808.6 110.2 698.4 0 562.5 0ZM562.5 328.1C517.2 328.1 480.5 291.4 480.5 246.1S517.2 164.1 562.5 164.1 644.5 200.8 644.5 246.1 607.8 328.1 562.5 328.1ZM39.3 421.8A62.5 62.5 0 0 0 0 479.8V968.7C0 990.8 22.3 1005.9 42.9 997.7L312.5 875V419.8C295.2 388.6 281.1 358.2 271 329.1L39.3 421.8ZM562.5 702.5C535 702.5 509 690.4 491.2 669.4 452.8 624 412 572.4 375 519.5V875L750 1000V519.5C713 572.4 672.2 624.1 633.8 669.4 616 690.4 590 702.5 562.5 702.5ZM1082.1 314.8L812.5 437.5V1000L1085.7 890.7A62.5 62.5 0 0 0 1125 832.7V343.8C1125 321.7 1102.7 306.6 1082.1 314.8Z",
+        "width": 1125
+      },
+      "search": [
+        "map-marked-alt"
+      ]
+    },
+    {
       "uid": "aaafc57606711978ce1817857229d8eb",
       "css": "plus",
       "code": 59500,
