+++ src/angular/planit/src/assets/fonts/fontello/css/temperate-codes.css
@@ -123,6 +123,7 @@
 .icon-mail:before { content: '\e879'; } /* '' */
 .icon-draw-polygon:before { content: '\e87a'; } /* '' */
 .icon-search:before { content: '\e87b'; } /* '' */
+.icon-map-marked-alt:before { content: '\e897'; } /* '' */
 .icon-forestry:before { content: '\e8e6'; } /* '' */
 .icon-circle:before { content: '\e8fc'; } /* '' */
 .icon-circle-empty:before { content: '\e8fd'; } /* '' */
