+++ src/angular/planit/src/assets/fonts/fontello/css/temperate.css
@@ -1,11 +1,11 @@
 @font-face {
   font-family: 'temperate';
-  src: url('../font/temperate.eot?96113663');
-  src: url('../font/temperate.eot?96113663#iefix') format('embedded-opentype'),
-       url('../font/temperate.woff2?96113663') format('woff2'),
-       url('../font/temperate.woff?96113663') format('woff'),
-       url('../font/temperate.ttf?96113663') format('truetype'),
-       url('../font/temperate.svg?96113663#temperate') format('svg');
+  src: url('../font/temperate.eot?65604603');
+  src: url('../font/temperate.eot?65604603#iefix') format('embedded-opentype'),
+       url('../font/temperate.woff2?65604603') format('woff2'),
+       url('../font/temperate.woff?65604603') format('woff'),
+       url('../font/temperate.ttf?65604603') format('truetype'),
+       url('../font/temperate.svg?65604603#temperate') format('svg');
   font-weight: normal;
   font-style: normal;
 }
@@ -15,7 +15,7 @@
 @media screen and (-webkit-min-device-pixel-ratio:0) {
   @font-face {
     font-family: 'temperate';
-    src: url('../font/temperate.svg?96113663#temperate') format('svg');
+    src: url('../font/temperate.svg?65604603#temperate') format('svg');
   }
 }
 */
@@ -179,6 +179,7 @@
 .icon-mail:before { content: '\e879'; } /* '' */
 .icon-draw-polygon:before { content: '\e87a'; } /* '' */
 .icon-search:before { content: '\e87b'; } /* '' */
+.icon-map-marked-alt:before { content: '\e897'; } /* '' */
 .icon-forestry:before { content: '\e8e6'; } /* '' */
 .icon-circle:before { content: '\e8fc'; } /* '' */
 .icon-circle-empty:before { content: '\e8fd'; } /* '' */
