+++ src/angular/planit/src/assets/fonts/fontello/demo.html
@@ -229,11 +229,11 @@ body {
 }
 @font-face {
       font-family: 'temperate';
-      src: url('./font/temperate.eot?51783189');
-      src: url('./font/temperate.eot?51783189#iefix') format('embedded-opentype'),
-           url('./font/temperate.woff?51783189') format('woff'),
-           url('./font/temperate.ttf?51783189') format('truetype'),
-           url('./font/temperate.svg?51783189#temperate') format('svg');
+      src: url('./font/temperate.eot?73398089');
+      src: url('./font/temperate.eot?73398089#iefix') format('embedded-opentype'),
+           url('./font/temperate.woff?73398089') format('woff'),
+           url('./font/temperate.ttf?73398089') format('truetype'),
+           url('./font/temperate.svg?73398089#temperate') format('svg');
       font-weight: normal;
       font-style: normal;
     }
@@ -484,18 +484,19 @@ body {
         <div class="the-icons span3" title="Code: 0xe87b"><i class="demo-icon icon-search">&#xe87b;</i> <span class="i-name">icon-search</span><span class="i-code">0xe87b</span></div>
       </div>
       <div class="row">
+        <div class="the-icons span3" title="Code: 0xe897"><i class="demo-icon icon-map-marked-alt">&#xe897;</i> <span class="i-name">icon-map-marked-alt</span><span class="i-code">0xe897</span></div>
         <div class="the-icons span3" title="Code: 0xe8e6"><i class="demo-icon icon-forestry">&#xe8e6;</i> <span class="i-name">icon-forestry</span><span class="i-code">0xe8e6</span></div>
         <div class="the-icons span3" title="Code: 0xe8fc"><i class="demo-icon icon-circle">&#xe8fc;</i> <span class="i-name">icon-circle</span><span class="i-code">0xe8fc</span></div>
         <div class="the-icons span3" title="Code: 0xe8fd"><i class="demo-icon icon-circle-empty">&#xe8fd;</i> <span class="i-name">icon-circle-empty</span><span class="i-code">0xe8fd</span></div>
-        <div class="the-icons span3" title="Code: 0xf099"><i class="demo-icon icon-twitter">&#xf099;</i> <span class="i-name">icon-twitter</span><span class="i-code">0xf099</span></div>
       </div>
       <div class="row">
+        <div class="the-icons span3" title="Code: 0xf099"><i class="demo-icon icon-twitter">&#xf099;</i> <span class="i-name">icon-twitter</span><span class="i-code">0xf099</span></div>
         <div class="the-icons span3" title="Code: 0xf09a"><i class="demo-icon icon-facebook">&#xf09a;</i> <span class="i-name">icon-facebook</span><span class="i-code">0xf09a</span></div>
         <div class="the-icons span3" title="Code: 0xf0d4"><i class="demo-icon icon-gplus-squared">&#xf0d4;</i> <span class="i-name">icon-gplus-squared</span><span class="i-code">0xf0d4</span></div>
         <div class="the-icons span3" title="Code: 0xf0d5"><i class="demo-icon icon-gplus">&#xf0d5;</i> <span class="i-name">icon-gplus</span><span class="i-code">0xf0d5</span></div>
-        <div class="the-icons span3" title="Code: 0xf230"><i class="demo-icon icon-facebook-official">&#xf230;</i> <span class="i-name">icon-facebook-official</span><span class="i-code">0xf230</span></div>
       </div>
       <div class="row">
+        <div class="the-icons span3" title="Code: 0xf230"><i class="demo-icon icon-facebook-official">&#xf230;</i> <span class="i-name">icon-facebook-official</span><span class="i-code">0xf230</span></div>
         <div class="the-icons span3" title="Code: 0xf308"><i class="demo-icon icon-facebook-squared">&#xf308;</i> <span class="i-name">icon-facebook-squared</span><span class="i-code">0xf308</span></div>
       </div>
     </div>
