+++ src/angular/planit/src/assets/sass/main.scss
@@ -53,6 +53,7 @@
   'components/marketing/faq',
   'components/marketing/marketing-cta',
   'components/shared/app-chart',
+  'components/shared/impact-map',
   'components/shared/ld-progressbar',
   'components/shared/option-dropdown',
   'components/shared/plan-cards',
@@ -102,7 +103,6 @@
 @import 'pages/indicators';
 @import 'pages/manage-subscription';
 @import 'pages/marketing';
-@import 'pages/map';
 @import 'pages/vulnerability-assessment';
 @import 'pages/risk-wizard';
 @import 'pages/action-wizard';
